csv파일 설정 배경 이미지 추가
1.농장 내부 플레이어 이동스크립트 제작 및 충돌처리
2.돼지 패널 생성시 앞을 가리는 문제 해결
3.애니매이션 수정 작업
4.pivot위치 수정
5.Dog스크립트 작성
6.개,알파카,소,양 스크립트 추가 작성
7.csv파일 설정 배경 이미지 추가

+++ RiotSample0/Assets/Scripts/PlayerInfo.cs
@@ -68,14 +68,14 @@ public class PlayerInfo : MonoBehaviour
     #endregion
 
     #region DPS
-    public void SetCharDps(int charID,int charDPS)
+    public void SetCharDps(int charID,float charDPS)
     {
-        PlayerPrefs.SetInt(charID + "DPS", charDPS);
+        PlayerPrefs.SetFloat(charID + "DPS", charDPS);
         UnityEngine.Debug.Log(charID + "DPS " + charDPS);
     }
-    public int GetCharDps(int charID)
+    public float GetCharDps(int charID)
     {
-        return PlayerPrefs.GetInt(charID + "DPS");
+        return PlayerPrefs.GetFloat(charID + "DPS");
     }
     #endregion
